requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
Pillow==10.2.0
cryptography==42.0.5
aiohttp==3.9.3